deepdream-generator-api/requirements.txt
python-lambda-local
pytest
boto3
pytest-dotenv
pynamodb
tensorflow
Pillow
deepdiff
